requirements.txt
asgiref==3.8.1
Django==5.1.2
django-debug-toolbar==4.4.6
psycopg==3.2.3
psycopg-c==3.2.3
psycopg-pool==3.2.4
sqlparse==0.5.1
typing_extensions==4.12.2
